agents/stage0_identity/pyproject.toml
[project]
name = "abunny-stage0-identity"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pipeline-contracts",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
abunny-stage0 = "abunny_stage0_identity.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/abunny_stage0_identity"]

[tool.uv.sources]
pipeline-contracts = { workspace = true }
